Add basic Elm app
This commit contains all the things which are needed to display
a simple "Hello, World!" in Elm in a Rails app.

The infrastructure for compiling the code is missing,
but it's not the point of focus as of now. The compilation step is going
to be solved later by using either elm-webpack-loader [1] or a Makefile.

[1] https://github.com/rtfeldman/elm-webpack-loader