Use webpack for compiling the Elm app
The main changes are:

1. The setup of the Elm app is no longer done in a file controlled
by Sprockets. More about this in the README.

2. There's no longer a separete folder with the Elm app and all its files.
elm-package.json lives in the root of the project while any current and future
Elm files should live in `app/assets/javascripts/elm`. The latter directory
is also added as to `source-directories` in elm-package.json.

3. The file required in the Rails app is now `order_form.webpack.js`.
More about this also in the README.